requirements.txt
argparse
matplotlib
nltk
numpy
Pillow
torch
torchvision
pycocotools
streamlit
streamlit-aggrid==0.2.2-2
-e .